Assessment/Backend/backend/api/serializers.py
from rest_framework import serializers
from .models import BlogPost, User


class BlogPostSerializer(serializers.ModelSerializer):
    class Meta:
        model = BlogPost
        fields = ["id", "title", "content", "published_date","authorid"]
    
class UserSerializer(serializers.ModelSerializer):
    class Meta:
        model = User
        fields = ["id","username","password","role"]