Assessment/Backend/requirements.txt
django
djangorestframework
environs
asgiref
django-cors-headers
djangorestframework-simplejwt
PyJWT
pytz
sqlparse
psycopg2-binary
python-dotenv